refactor: restructure and move stuff, form public API
